Fixes for Debug build.

+++ DGGraphImpl.cpp
@@ -1869,14 +1869,14 @@ void DGGraphImpl::loadKLOperatorSourceCode(const std::string & name, const std::
   }
 
   fseek( file, 0, SEEK_END );
-  int fileSize = ftell( file );
+  long fileSize = ftell( file );
   rewind( file );
 
   char * buffer = (char*) malloc(fileSize + 1);
   buffer[fileSize] = '\0';
 
   size_t readBytes = fread(buffer, 1, fileSize, file);
-  assert(readBytes == fileSize);
+  assert(readBytes == size_t(fileSize));
   (void)readBytes;
 
   fclose(file);
@@ -3060,14 +3060,14 @@ bool DGGraphImpl::loadFromFile(
     return LoggingImpl::reportError("Invalid filePath '"+filePath+"'", errorOut);
 
   fseek( file, 0, SEEK_END );
-  int fileSize = ftell( file );
+  long fileSize = ftell( file );
   rewind( file );
 
   char * buffer = (char*) malloc(fileSize + 1);
   buffer[fileSize] = '\0';
 
   size_t readBytes = fread(buffer, 1, fileSize, file);
-  assert(readBytes == fileSize);
+  assert(readBytes == size_t(fileSize));
   (void)readBytes;
 
   fclose(file);
